refactor: Car 클래스 분리, UI 로직과 비즈니스 로직 분리
