+++ src/main/kotlin/racingcar/PrintCarPosition.kt
@@ -1,0 +1,14 @@
+package racingcar
+
+const val CAR_CHARACTER = '-'
+
+fun printDefaultPosition() {
+    println(CAR_CHARACTER)
+}
+
+fun printCurrentPosition(position: Int) {
+    repeat(position) {
+        print(CAR_CHARACTER)
+    }
+    println()
+}
